throw if user not creator

+++ packages/liveblocks-react/src/room.tsx
@@ -1400,6 +1400,20 @@ function useDeleteThread(): (threadId: string) => void {
       const optimisticUpdateId = nanoid();
 
       const { store, onMutationFailure } = getExtrasForClient(client);
+
+      const thread = store.get().threads[threadId];
+
+      const userId = getCurrentUserId(room);
+
+      if (
+        thread &&
+        thread.comments &&
+        thread.comments[0] &&
+        thread.comments[0].userId !== userId
+      ) {
+        throw new Error("Only the thread creator can delete the thread");
+      }
+
       store.pushOptimisticUpdate({
         type: "delete-thread",
         id: optimisticUpdateId,
